did: merge very small files, make use of new htmlgenerator lazy attribute accessor
