--- bread/layout/components/forms/email_input.py
@@ -1,6 +1,0 @@
-from .text_input import TextInput
-
-
-class EmailInput(TextInput):
-    def __init__(self, **attributes):
-        super().__init__(icon="email", **attributes)
