--- bread/layout/components/forms/phone_number_input.py
@@ -1,6 +1,0 @@
-from .text_input import TextInput
-
-
-class PhoneNumberInput(TextInput):
-    def __init__(self, **attributes):
-        super().__init__(icon="phone", **attributes)
