--- bread/layout/components/forms/url_input.py
@@ -1,6 +1,0 @@
-from .text_input import TextInput
-
-
-class UrlInput(TextInput):
-    def __init__(self, **attributes):
-        super().__init__(icon="link", **attributes)
